Updated collision, added SoundManager
Circle intersecting with a square is now getting pushed out of square if
player is in,
SoundManager plays music, added several audio files.

+++ BurstinBubbles/Enemy.h
@@ -17,10 +17,10 @@ public:
 	virtual void Die(void);
 	std::vector<Bullet*> m_bullets;
 
-	static float fMAX_ACTIVE_ENEMIES;
-	static float fCurrentEnemies;
-	static float fTOTAL_ENEMIES_PER_SCENE;
-	static float fSPAWNED_ENEMIES_PER_SCENE;
+	static int fMAX_ACTIVE_ENEMIES;
+	static int fCurrentEnemies;
+	static int fTOTAL_ENEMIES_PER_SCENE;
+	static int fSPAWNED_ENEMIES_PER_SCENE;
 private:
 	float m_fShootTimer;
 };
